+++ public/sw.js
@@ -1,1 +1,1 @@
-if(!self.define){let e,s={};const c=(c,n)=>(c=new URL(c+".js",n).href,s[c]||new Promise((s=>{if("document"in self){const e=document.createElement("script");e.src=c,e.onload=s,document.head.appendChild(e)}else e=c,importScripts(c),s()})).then((()=>{let e=s[c];if(!e)throw new Error(`Module ${c} didn’t register its module`);return e})));self.define=(n,i)=>{const a=e||("document"in self?document.currentScript.src:"")||location.href;if(s[a])return;let t={};const r=e=>c(e,a),d={module:{uri:a},exports:t,require:r};s[a]=Promise.all(n.map((e=>d[e]||r(e)))).then((e=>(i(...e),t)))}}define(["./workbox-6316bd60"],(function(e){"use strict";importScripts("fallback-lUgri-bKWHtk5a7SsRDE2.js"),self.skipWaiting(),e.clientsClaim(),e.precacheAndRoute([{url:"/_next/static/chunks/1046.f9920292acb886cf.js",revision:"f9920292acb886cf"},{url:"/_next/static/chunks/1136.5cfda1c217a53465.js",revision:"5cfda1c217a53465"},{url:"/_next/static/chunks/1172.fcf084343ac422c0.js",revision:"fcf084343ac422c0"},{url:"/_next/static/chunks/139-3b2f902655379be7.js",revision:"3b2f902655379be7"},{url:"/_next/static/chunks/1606726a.7708375a5afc05cc.js",revision:"7708375a5afc05cc"},{url:"/_next/static/chunks/1630.1304d27e48335005.js",revision:"1304d27e48335005"},{url:"/_next/static/chunks/1881.7339b86b2652ea41.js",revision:"7339b86b2652ea41"},{url:"/_next/static/chunks/29107295-4a69275373f23f88.js",revision:"4a69275373f23f88"},{url:"/_next/static/chunks/3281.183a00530cd049ff.js",revision:"183a00530cd049ff"},{url:"/_next/static/chunks/3531-9a7ab27147f46cd4.js",revision:"9a7ab27147f46cd4"},{url:"/_next/static/chunks/373-66085ded0614f6e9.js",revision:"66085ded0614f6e9"},{url:"/_next/static/chunks/3929.768e7d7ce578cccd.js",revision:"768e7d7ce578cccd"},{url:"/_next/static/chunks/415-a7041e0cd73f4d02.js",revision:"a7041e0cd73f4d02"},{url:"/_next/static/chunks/4845.9059dab29e7f39e0.js",revision:"9059dab29e7f39e0"},{url:"/_next/static/chunks/4971-bf610639b07776f0.js",revision:"bf610639b07776f0"},{url:"/_next/static/chunks/5054-46838bbc5293f777.js",revision:"46838bbc5293f777"},{url:"/_next/static/chunks/5132.5d3cdca0333210e1.js",revision:"5d3cdca0333210e1"},{url:"/_next/static/chunks/55a21ef8.ce7a778c6b500e34.js",revision:"ce7a778c6b500e34"},{url:"/_next/static/chunks/5981.14861835b1a1e818.js",revision:"14861835b1a1e818"},{url:"/_next/static/chunks/6250-afb82d45d1b1478b.js",revision:"afb82d45d1b1478b"},{url:"/_next/static/chunks/6548.25eb4110c2af94fb.js",revision:"25eb4110c2af94fb"},{url:"/_next/static/chunks/6806.1baf9f26569ea39a.js",revision:"1baf9f26569ea39a"},{url:"/_next/static/chunks/6807.a346c5bfbdc47385.js",revision:"a346c5bfbdc47385"},{url:"/_next/static/chunks/7329-c7483d0bf284e45a.js",revision:"c7483d0bf284e45a"},{url:"/_next/static/chunks/7452.7c56c8317332f457.js",revision:"7c56c8317332f457"},{url:"/_next/static/chunks/7665-88ce5d4405bcc3a8.js",revision:"88ce5d4405bcc3a8"},{url:"/_next/static/chunks/7674.314782bedd2e0ef5.js",revision:"314782bedd2e0ef5"},{url:"/_next/static/chunks/7757-fe219f6fb03eb508.js",revision:"fe219f6fb03eb508"},{url:"/_next/static/chunks/7779.2926b1770bde045e.js",revision:"2926b1770bde045e"},{url:"/_next/static/chunks/7847-e4808383e9667a22.js",revision:"e4808383e9667a22"},{url:"/_next/static/chunks/8090.8d68f5dfff066c68.js",revision:"8d68f5dfff066c68"},{url:"/_next/static/chunks/8142-f2ed631727cc671c.js",revision:"f2ed631727cc671c"},{url:"/_next/static/chunks/8254.a7059c867670d934.js",revision:"a7059c867670d934"},{url:"/_next/static/chunks/8282.4f4796ba81b6b048.js",revision:"4f4796ba81b6b048"},{url:"/_next/static/chunks/8526-cd92e4c33d0bbf1a.js",revision:"cd92e4c33d0bbf1a"},{url:"/_next/static/chunks/9138.b014005ebc70373a.js",revision:"b014005ebc70373a"},{url:"/_next/static/chunks/9382c5f8.6fa4d46c1a7ce68c.js",revision:"6fa4d46c1a7ce68c"},{url:"/_next/static/chunks/950.2667035e16942e75.js",revision:"2667035e16942e75"},{url:"/_next/static/chunks/9536.19731d6d9fbac332.js",revision:"19731d6d9fbac332"},{url:"/_next/static/chunks/9758.8884e9bdf2346f4b.js",revision:"8884e9bdf2346f4b"},{url:"/_next/static/chunks/9792.696c67dff6b73dd3.js",revision:"696c67dff6b73dd3"},{url:"/_next/static/chunks/983.88c32f7bad8bfb89.js",revision:"88c32f7bad8bfb89"},{url:"/_next/static/chunks/9878.dd875d173a13fbf6.js",revision:"dd875d173a13fbf6"},{url:"/_next/static/chunks/ccf7da5b.2263ee63469cf75b.js",revision:"2263ee63469cf75b"},{url:"/_next/static/chunks/de49cc29.7667da28d63bf787.js",revision:"7667da28d63bf787"},{url:"/_next/static/chunks/fec483df.6b764e0fbcca117e.js",revision:"6b764e0fbcca117e"},{url:"/_next/static/chunks/framework-79bce4a3a540b080.js",revision:"79bce4a3a540b080"},{url:"/_next/static/chunks/main-8bdc27600901d989.js",revision:"8bdc27600901d989"},{url:"/_next/static/chunks/pages/_app-643af7efcd7b5a0d.js",revision:"643af7efcd7b5a0d"},{url:"/_next/static/chunks/pages/_error-d39607a4676a4aa5.js",revision:"d39607a4676a4aa5"},{url:"/_next/static/chunks/pages/_home-1468f0ac001adae6.js",revision:"1468f0ac001adae6"},{url:"/_next/static/chunks/pages/_offline-78a77348806b3a4a.js",revision:"78a77348806b3a4a"},{url:"/_next/static/chunks/pages/comment-policy-daf8a341459b849b.js",revision:"daf8a341459b849b"},{url:"/_next/static/chunks/pages/dashboard-e3065690f1c21170.js",revision:"e3065690f1c21170"},{url:"/_next/static/chunks/pages/game/%5Bid%5D-c31c6e70ed421501.js",revision:"c31c6e70ed421501"},{url:"/_next/static/chunks/pages/game/edit/%5Bid%5D-f74b3c1b80bb8465.js",revision:"f74b3c1b80bb8465"},{url:"/_next/static/chunks/pages/game/new-4bb9e389f1569ecd.js",revision:"4bb9e389f1569ecd"},{url:"/_next/static/chunks/pages/games-6cbfbdc16cc9aea2.js",revision:"6cbfbdc16cc9aea2"},{url:"/_next/static/chunks/pages/index-edd03e78bf84a4c7.js",revision:"edd03e78bf84a4c7"},{url:"/_next/static/chunks/pages/login-14e90500fdd88255.js",revision:"14e90500fdd88255"},{url:"/_next/static/chunks/pages/logout-0bdb809bfc47f2a9.js",revision:"0bdb809bfc47f2a9"},{url:"/_next/static/chunks/pages/profile/%5Busername%5D-f87dc5585e403b6a.js",revision:"f87dc5585e403b6a"},{url:"/_next/static/chunks/pages/profile/_home-bccf2f059b60a934.js",revision:"bccf2f059b60a934"},{url:"/_next/static/chunks/pages/register-cee6796b6878021d.js",revision:"cee6796b6878021d"},{url:"/_next/static/chunks/pages/user/settings-699d7273f70a5c0d.js",revision:"699d7273f70a5c0d"},{url:"/_next/static/chunks/pages/user/settings/_layout-c1b5091c29e6e151.js",revision:"c1b5091c29e6e151"},{url:"/_next/static/chunks/pages/user/settings/analytics-836511ad46dddc81.js",revision:"836511ad46dddc81"},{url:"/_next/static/chunks/pages/user/settings/api-keys-a71521607fbb01a3.js",revision:"a71521607fbb01a3"},{url:"/_next/static/chunks/pages/user/settings/connected-accounts-22e21d56ba8dafc2.js",revision:"22e21d56ba8dafc2"},{url:"/_next/static/chunks/pages/user/settings/data-export-812ce98177cca3cd.js",revision:"812ce98177cca3cd"},{url:"/_next/static/chunks/pages/user/settings/oauth-apps-447f0943ebab6dc1.js",revision:"447f0943ebab6dc1"},{url:"/_next/static/chunks/pages/user/settings/press-cefb15e1f334a82f.js",revision:"cefb15e1f334a82f"},{url:"/_next/static/chunks/pages/user/settings/privacy-eee940413dffb980.js",revision:"eee940413dffb980"},{url:"/_next/static/chunks/polyfills-5cd94c89d3acac5f.js",revision:"99442aec5788bccac9b2f0ead2afdd6b"},{url:"/_next/static/chunks/webpack-f8d9c8fae778832d.js",revision:"f8d9c8fae778832d"},{url:"/_next/static/css/027e2130b268b60e.css",revision:"027e2130b268b60e"},{url:"/_next/static/css/0b5cbdb6fcbeeba6.css",revision:"0b5cbdb6fcbeeba6"},{url:"/_next/static/css/0d8da352da0c482e.css",revision:"0d8da352da0c482e"},{url:"/_next/static/css/13e43ffa46de2cdd.css",revision:"13e43ffa46de2cdd"},{url:"/_next/static/css/149b18973e5508c7.css",revision:"149b18973e5508c7"},{url:"/_next/static/css/7fad8ef8758b2544.css",revision:"7fad8ef8758b2544"},{url:"/_next/static/lUgri-bKWHtk5a7SsRDE2/_buildManifest.js",revision:"36391e48aa161d9fe2d123f7709b2abe"},{url:"/_next/static/lUgri-bKWHtk5a7SsRDE2/_middlewareManifest.js",revision:"fb2823d66b3e778e04a3f681d0d2fb19"},{url:"/_next/static/lUgri-bKWHtk5a7SsRDE2/_ssgManifest.js",revision:"b6652df95db52feb4daf4eca35380933"},{url:"/_offline",revision:"lUgri-bKWHtk5a7SsRDE2"},{url:"/android-chrome-192x192.png",revision:"94d6bfe86565eb36002c36e3fb5c0fec"},{url:"/android-chrome-384x384.png",revision:"85cdc32664335342102d91009859e328"},{url:"/android-chrome-512x512.png",revision:"f1112f7e13556d584083088dd1fef001"},{url:"/apple-touch-icon.png",revision:"a73b67e4e500429f2915cd7f77248ac3"},{url:"/favicon-16x16.png",revision:"72b83ca0dc00bbd87324e8142994d2f5"},{url:"/favicon-32x32.png",revision:"72273e0862fa09db7e6018a5c0f21700"},{url:"/favicon.ico",revision:"35738d3356ffcab9c365dab32c88bf18"},{url:"/icons/android-chrome-192x192.png",revision:"e866b081cb419c9ff5c65648878f4929"},{url:"/icons/android-chrome-384x384.png",revision:"85cdc32664335342102d91009859e328"},{url:"/icons/arbitrum-logo.svg",revision:"b0ee7da543ad178259fd73d2db2bb5e5"},{url:"/icons/binance-bnb-logo.svg",revision:"1aa88810b9f05826b0ce034c0799f7e6"},{url:"/icons/ethereum-eth-logo.svg",revision:"7c19954628081bf6fd772fa0d0819547"},{url:"/icons/fonts/icomoon.eot",revision:"b89c270dc501f53c325c413d1a3e709f"},{url:"/icons/fonts/icomoon.svg",revision:"7c6ee2b00874d8b175719bf35b54a0ba"},{url:"/icons/fonts/icomoon.ttf",revision:"14ab7dc517b08e9e2f67e0b7288d79e6"},{url:"/icons/fonts/icomoon.woff",revision:"edbdd69608a25fad5bfcd861f59c9fee"},{url:"/icons/icon-512x512.png",revision:"0f1c149d1908755365d1b788885da076"},{url:"/icons/metamask.svg",revision:"925c6160dccb471920ca5cbe265b3859"},{url:"/icons/optimistic-ethereum-logo.svg",revision:"15863d469e2ea202ed0f4adc542cd142"},{url:"/icons/polygon-matic-logo.svg",revision:"1087d3d671ac2ca5eeea49c18e1175bf"},{url:"/icons/sort.svg",revision:"1bea8d5724ad07940bc563a5b7896ae4"},{url:"/icons/tag.svg",revision:"6a048381eca9ed5285eae2b7ab3d9708"},{url:"/icons/walletconnect.svg",revision:"8ec0ff1250bda2b0a5548528310cc55a"},{url:"/images/loader.gif",revision:"a686deb090f7f8d412cf5faf6c06f4cd"},{url:"/images/oi9XNx.png",revision:"6573ce0fbc54f71da2bd39ece24806bc"},{url:"/manifest.json",revision:"7bcdc3a803995327c9f507fc30b9004d"},{url:"/touch-icon-ipad-retina.png",revision:"a73b67e4e500429f2915cd7f77248ac3"},{url:"/touch-icon-ipad.png",revision:"a73b67e4e500429f2915cd7f77248ac3"},{url:"/touch-icon-iphone-retina.png",revision:"a73b67e4e500429f2915cd7f77248ac3"},{url:"/touch-icon-iphone.png",revision:"a73b67e4e500429f2915cd7f77248ac3"},{url:"/vercel.svg",revision:"26bf2d0adaf1028a4d4c6ee77005e819"}],{ignoreURLParametersMatching:[]}),e.cleanupOutdatedCaches(),e.registerRoute("/",new e.NetworkFirst({cacheName:"start-url",plugins:[{cacheWillUpdate:async({request:e,response:s,event:c,state:n})=>s&&"opaqueredirect"===s.type?new Response(s.body,{status:200,statusText:"OK",headers:s.headers}):s},{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:gstatic)\.com\/.*/i,new e.CacheFirst({cacheName:"google-fonts-webfonts",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:31536e3}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:googleapis)\.com\/.*/i,new e.StaleWhileRevalidate({cacheName:"google-fonts-stylesheets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\.(?:eot|otf|ttc|ttf|woff|woff2|font.css)$/i,new e.StaleWhileRevalidate({cacheName:"static-font-assets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\.(?:jpg|jpeg|gif|png|svg|ico|webp)$/i,new e.StaleWhileRevalidate({cacheName:"static-image-assets",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\/_next\/image\?url=.+$/i,new e.StaleWhileRevalidate({cacheName:"next-image",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\.(?:mp3|wav|ogg)$/i,new e.CacheFirst({cacheName:"static-audio-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\.(?:mp4)$/i,new e.CacheFirst({cacheName:"static-video-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\.(?:js)$/i,new e.StaleWhileRevalidate({cacheName:"static-js-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\.(?:css|less)$/i,new e.StaleWhileRevalidate({cacheName:"static-style-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\/_next\/data\/.+\/.+\.json$/i,new e.StaleWhileRevalidate({cacheName:"next-data",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\.(?:json|xml|csv)$/i,new e.NetworkFirst({cacheName:"static-data-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute((({url:e})=>{if(!(self.origin===e.origin))return!1;const s=e.pathname;return!s.startsWith("/api/auth/")&&!!s.startsWith("/api/")}),new e.NetworkFirst({cacheName:"apis",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:16,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute((({url:e})=>{if(!(self.origin===e.origin))return!1;return!e.pathname.startsWith("/api/")}),new e.NetworkFirst({cacheName:"others",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute((({url:e})=>!(self.origin===e.origin)),new e.NetworkFirst({cacheName:"cross-origin",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:3600}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET")}));
+if(!self.define){let e,s={};const c=(c,n)=>(c=new URL(c+".js",n).href,s[c]||new Promise((s=>{if("document"in self){const e=document.createElement("script");e.src=c,e.onload=s,document.head.appendChild(e)}else e=c,importScripts(c),s()})).then((()=>{let e=s[c];if(!e)throw new Error(`Module ${c} didn’t register its module`);return e})));self.define=(n,a)=>{const i=e||("document"in self?document.currentScript.src:"")||location.href;if(s[i])return;let t={};const r=e=>c(e,i),d={module:{uri:i},exports:t,require:r};s[i]=Promise.all(n.map((e=>d[e]||r(e)))).then((e=>(a(...e),t)))}}define(["./workbox-6316bd60"],(function(e){"use strict";importScripts("fallback-gwPJsm4GfjPgra8LQQOM8.js"),self.skipWaiting(),e.clientsClaim(),e.precacheAndRoute([{url:"/_next/static/chunks/1046.f9920292acb886cf.js",revision:"f9920292acb886cf"},{url:"/_next/static/chunks/1136.5cfda1c217a53465.js",revision:"5cfda1c217a53465"},{url:"/_next/static/chunks/1172.fcf084343ac422c0.js",revision:"fcf084343ac422c0"},{url:"/_next/static/chunks/123-811cc3d1961c82c3.js",revision:"811cc3d1961c82c3"},{url:"/_next/static/chunks/139-3b2f902655379be7.js",revision:"3b2f902655379be7"},{url:"/_next/static/chunks/1562-4566afba394d7640.js",revision:"4566afba394d7640"},{url:"/_next/static/chunks/1606726a.7708375a5afc05cc.js",revision:"7708375a5afc05cc"},{url:"/_next/static/chunks/1630.1304d27e48335005.js",revision:"1304d27e48335005"},{url:"/_next/static/chunks/1881.7339b86b2652ea41.js",revision:"7339b86b2652ea41"},{url:"/_next/static/chunks/29107295-4a69275373f23f88.js",revision:"4a69275373f23f88"},{url:"/_next/static/chunks/3281.183a00530cd049ff.js",revision:"183a00530cd049ff"},{url:"/_next/static/chunks/3531-9a7ab27147f46cd4.js",revision:"9a7ab27147f46cd4"},{url:"/_next/static/chunks/373-2d658c2b96df90ff.js",revision:"2d658c2b96df90ff"},{url:"/_next/static/chunks/3929.768e7d7ce578cccd.js",revision:"768e7d7ce578cccd"},{url:"/_next/static/chunks/4845.9059dab29e7f39e0.js",revision:"9059dab29e7f39e0"},{url:"/_next/static/chunks/5054-46838bbc5293f777.js",revision:"46838bbc5293f777"},{url:"/_next/static/chunks/5132.5d3cdca0333210e1.js",revision:"5d3cdca0333210e1"},{url:"/_next/static/chunks/55a21ef8.ce7a778c6b500e34.js",revision:"ce7a778c6b500e34"},{url:"/_next/static/chunks/5981.14861835b1a1e818.js",revision:"14861835b1a1e818"},{url:"/_next/static/chunks/6250-8275c4ab867c6769.js",revision:"8275c4ab867c6769"},{url:"/_next/static/chunks/6548.25eb4110c2af94fb.js",revision:"25eb4110c2af94fb"},{url:"/_next/static/chunks/6806.1baf9f26569ea39a.js",revision:"1baf9f26569ea39a"},{url:"/_next/static/chunks/6807.a346c5bfbdc47385.js",revision:"a346c5bfbdc47385"},{url:"/_next/static/chunks/7329-205dc3418f6e3a96.js",revision:"205dc3418f6e3a96"},{url:"/_next/static/chunks/7452.7c56c8317332f457.js",revision:"7c56c8317332f457"},{url:"/_next/static/chunks/7665-88ce5d4405bcc3a8.js",revision:"88ce5d4405bcc3a8"},{url:"/_next/static/chunks/7674.314782bedd2e0ef5.js",revision:"314782bedd2e0ef5"},{url:"/_next/static/chunks/7779.2926b1770bde045e.js",revision:"2926b1770bde045e"},{url:"/_next/static/chunks/8073-7252eb31631e74d1.js",revision:"7252eb31631e74d1"},{url:"/_next/static/chunks/8090.8d68f5dfff066c68.js",revision:"8d68f5dfff066c68"},{url:"/_next/static/chunks/8142-09f3b8d335c16c20.js",revision:"09f3b8d335c16c20"},{url:"/_next/static/chunks/8254.a7059c867670d934.js",revision:"a7059c867670d934"},{url:"/_next/static/chunks/8282.4f4796ba81b6b048.js",revision:"4f4796ba81b6b048"},{url:"/_next/static/chunks/8526-34cee8239e187f37.js",revision:"34cee8239e187f37"},{url:"/_next/static/chunks/9138.b014005ebc70373a.js",revision:"b014005ebc70373a"},{url:"/_next/static/chunks/9382c5f8.6fa4d46c1a7ce68c.js",revision:"6fa4d46c1a7ce68c"},{url:"/_next/static/chunks/950.2667035e16942e75.js",revision:"2667035e16942e75"},{url:"/_next/static/chunks/9536.19731d6d9fbac332.js",revision:"19731d6d9fbac332"},{url:"/_next/static/chunks/9758.8884e9bdf2346f4b.js",revision:"8884e9bdf2346f4b"},{url:"/_next/static/chunks/9792.696c67dff6b73dd3.js",revision:"696c67dff6b73dd3"},{url:"/_next/static/chunks/983.88c32f7bad8bfb89.js",revision:"88c32f7bad8bfb89"},{url:"/_next/static/chunks/9878.dd875d173a13fbf6.js",revision:"dd875d173a13fbf6"},{url:"/_next/static/chunks/ccf7da5b.2263ee63469cf75b.js",revision:"2263ee63469cf75b"},{url:"/_next/static/chunks/de49cc29.7667da28d63bf787.js",revision:"7667da28d63bf787"},{url:"/_next/static/chunks/fec483df.6b764e0fbcca117e.js",revision:"6b764e0fbcca117e"},{url:"/_next/static/chunks/framework-79bce4a3a540b080.js",revision:"79bce4a3a540b080"},{url:"/_next/static/chunks/main-8bdc27600901d989.js",revision:"8bdc27600901d989"},{url:"/_next/static/chunks/pages/_app-95cb8cff49f21591.js",revision:"95cb8cff49f21591"},{url:"/_next/static/chunks/pages/_error-d39607a4676a4aa5.js",revision:"d39607a4676a4aa5"},{url:"/_next/static/chunks/pages/_home-1468f0ac001adae6.js",revision:"1468f0ac001adae6"},{url:"/_next/static/chunks/pages/_offline-78a77348806b3a4a.js",revision:"78a77348806b3a4a"},{url:"/_next/static/chunks/pages/comment-policy-daf8a341459b849b.js",revision:"daf8a341459b849b"},{url:"/_next/static/chunks/pages/dashboard-0f2e71161b05ce10.js",revision:"0f2e71161b05ce10"},{url:"/_next/static/chunks/pages/game/%5Bid%5D-d5ea82b09d0bc6ae.js",revision:"d5ea82b09d0bc6ae"},{url:"/_next/static/chunks/pages/game/edit/%5Bid%5D-a4746daaafe55fab.js",revision:"a4746daaafe55fab"},{url:"/_next/static/chunks/pages/game/new-7a71bfdcb96e07ec.js",revision:"7a71bfdcb96e07ec"},{url:"/_next/static/chunks/pages/games-41513f3cf9cb1142.js",revision:"41513f3cf9cb1142"},{url:"/_next/static/chunks/pages/index-edd03e78bf84a4c7.js",revision:"edd03e78bf84a4c7"},{url:"/_next/static/chunks/pages/login-14e90500fdd88255.js",revision:"14e90500fdd88255"},{url:"/_next/static/chunks/pages/logout-0bdb809bfc47f2a9.js",revision:"0bdb809bfc47f2a9"},{url:"/_next/static/chunks/pages/profile/%5Busername%5D-f87dc5585e403b6a.js",revision:"f87dc5585e403b6a"},{url:"/_next/static/chunks/pages/profile/_home-bccf2f059b60a934.js",revision:"bccf2f059b60a934"},{url:"/_next/static/chunks/pages/register-cee6796b6878021d.js",revision:"cee6796b6878021d"},{url:"/_next/static/chunks/pages/user/settings-4f8e9df26116a52e.js",revision:"4f8e9df26116a52e"},{url:"/_next/static/chunks/pages/user/settings/_layout-02a94c7fc8ca97df.js",revision:"02a94c7fc8ca97df"},{url:"/_next/static/chunks/pages/user/settings/analytics-15035016c86e6082.js",revision:"15035016c86e6082"},{url:"/_next/static/chunks/pages/user/settings/api-keys-b4468cdd00fe89bd.js",revision:"b4468cdd00fe89bd"},{url:"/_next/static/chunks/pages/user/settings/connected-accounts-e0799e2beb19110e.js",revision:"e0799e2beb19110e"},{url:"/_next/static/chunks/pages/user/settings/data-export-0f7e968c78ff2009.js",revision:"0f7e968c78ff2009"},{url:"/_next/static/chunks/pages/user/settings/oauth-apps-441b77e4fb8564cb.js",revision:"441b77e4fb8564cb"},{url:"/_next/static/chunks/pages/user/settings/press-b618c0ee541e429c.js",revision:"b618c0ee541e429c"},{url:"/_next/static/chunks/pages/user/settings/privacy-cbe82938aae40bac.js",revision:"cbe82938aae40bac"},{url:"/_next/static/chunks/polyfills-5cd94c89d3acac5f.js",revision:"99442aec5788bccac9b2f0ead2afdd6b"},{url:"/_next/static/chunks/webpack-02cad0b29bc89768.js",revision:"02cad0b29bc89768"},{url:"/_next/static/css/027e2130b268b60e.css",revision:"027e2130b268b60e"},{url:"/_next/static/css/0d8da352da0c482e.css",revision:"0d8da352da0c482e"},{url:"/_next/static/css/13e43ffa46de2cdd.css",revision:"13e43ffa46de2cdd"},{url:"/_next/static/css/149b18973e5508c7.css",revision:"149b18973e5508c7"},{url:"/_next/static/css/7fad8ef8758b2544.css",revision:"7fad8ef8758b2544"},{url:"/_next/static/css/b5ded84ad0878756.css",revision:"b5ded84ad0878756"},{url:"/_next/static/gwPJsm4GfjPgra8LQQOM8/_buildManifest.js",revision:"b5d7aba687672b273d84fd8b9fd00d6c"},{url:"/_next/static/gwPJsm4GfjPgra8LQQOM8/_middlewareManifest.js",revision:"fb2823d66b3e778e04a3f681d0d2fb19"},{url:"/_next/static/gwPJsm4GfjPgra8LQQOM8/_ssgManifest.js",revision:"b6652df95db52feb4daf4eca35380933"},{url:"/_offline",revision:"gwPJsm4GfjPgra8LQQOM8"},{url:"/android-chrome-192x192.png",revision:"94d6bfe86565eb36002c36e3fb5c0fec"},{url:"/android-chrome-384x384.png",revision:"85cdc32664335342102d91009859e328"},{url:"/android-chrome-512x512.png",revision:"f1112f7e13556d584083088dd1fef001"},{url:"/apple-touch-icon.png",revision:"a73b67e4e500429f2915cd7f77248ac3"},{url:"/favicon-16x16.png",revision:"72b83ca0dc00bbd87324e8142994d2f5"},{url:"/favicon-32x32.png",revision:"72273e0862fa09db7e6018a5c0f21700"},{url:"/favicon.ico",revision:"35738d3356ffcab9c365dab32c88bf18"},{url:"/icons/android-chrome-192x192.png",revision:"e866b081cb419c9ff5c65648878f4929"},{url:"/icons/android-chrome-384x384.png",revision:"85cdc32664335342102d91009859e328"},{url:"/icons/arbitrum-logo.svg",revision:"b0ee7da543ad178259fd73d2db2bb5e5"},{url:"/icons/binance-bnb-logo.svg",revision:"1aa88810b9f05826b0ce034c0799f7e6"},{url:"/icons/ethereum-eth-logo.svg",revision:"7c19954628081bf6fd772fa0d0819547"},{url:"/icons/fonts/icomoon.eot",revision:"b89c270dc501f53c325c413d1a3e709f"},{url:"/icons/fonts/icomoon.svg",revision:"7c6ee2b00874d8b175719bf35b54a0ba"},{url:"/icons/fonts/icomoon.ttf",revision:"14ab7dc517b08e9e2f67e0b7288d79e6"},{url:"/icons/fonts/icomoon.woff",revision:"edbdd69608a25fad5bfcd861f59c9fee"},{url:"/icons/icon-512x512.png",revision:"0f1c149d1908755365d1b788885da076"},{url:"/icons/metamask.svg",revision:"925c6160dccb471920ca5cbe265b3859"},{url:"/icons/optimistic-ethereum-logo.svg",revision:"15863d469e2ea202ed0f4adc542cd142"},{url:"/icons/polygon-matic-logo.svg",revision:"1087d3d671ac2ca5eeea49c18e1175bf"},{url:"/icons/sort.svg",revision:"1bea8d5724ad07940bc563a5b7896ae4"},{url:"/icons/tag.svg",revision:"6a048381eca9ed5285eae2b7ab3d9708"},{url:"/icons/walletconnect.svg",revision:"8ec0ff1250bda2b0a5548528310cc55a"},{url:"/images/loader.gif",revision:"a686deb090f7f8d412cf5faf6c06f4cd"},{url:"/images/oi9XNx.png",revision:"6573ce0fbc54f71da2bd39ece24806bc"},{url:"/logo.png",revision:"f1112f7e13556d584083088dd1fef001"},{url:"/manifest.json",revision:"7bcdc3a803995327c9f507fc30b9004d"},{url:"/touch-icon-ipad-retina.png",revision:"a73b67e4e500429f2915cd7f77248ac3"},{url:"/touch-icon-ipad.png",revision:"a73b67e4e500429f2915cd7f77248ac3"},{url:"/touch-icon-iphone-retina.png",revision:"a73b67e4e500429f2915cd7f77248ac3"},{url:"/touch-icon-iphone.png",revision:"a73b67e4e500429f2915cd7f77248ac3"},{url:"/vercel.svg",revision:"26bf2d0adaf1028a4d4c6ee77005e819"}],{ignoreURLParametersMatching:[]}),e.cleanupOutdatedCaches(),e.registerRoute("/",new e.NetworkFirst({cacheName:"start-url",plugins:[{cacheWillUpdate:async({request:e,response:s,event:c,state:n})=>s&&"opaqueredirect"===s.type?new Response(s.body,{status:200,statusText:"OK",headers:s.headers}):s},{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:gstatic)\.com\/.*/i,new e.CacheFirst({cacheName:"google-fonts-webfonts",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:31536e3}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:googleapis)\.com\/.*/i,new e.StaleWhileRevalidate({cacheName:"google-fonts-stylesheets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\.(?:eot|otf|ttc|ttf|woff|woff2|font.css)$/i,new e.StaleWhileRevalidate({cacheName:"static-font-assets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\.(?:jpg|jpeg|gif|png|svg|ico|webp)$/i,new e.StaleWhileRevalidate({cacheName:"static-image-assets",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\/_next\/image\?url=.+$/i,new e.StaleWhileRevalidate({cacheName:"next-image",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\.(?:mp3|wav|ogg)$/i,new e.CacheFirst({cacheName:"static-audio-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\.(?:mp4)$/i,new e.CacheFirst({cacheName:"static-video-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\.(?:js)$/i,new e.StaleWhileRevalidate({cacheName:"static-js-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\.(?:css|less)$/i,new e.StaleWhileRevalidate({cacheName:"static-style-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\/_next\/data\/.+\/.+\.json$/i,new e.StaleWhileRevalidate({cacheName:"next-data",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\.(?:json|xml|csv)$/i,new e.NetworkFirst({cacheName:"static-data-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute((({url:e})=>{if(!(self.origin===e.origin))return!1;const s=e.pathname;return!s.startsWith("/api/auth/")&&!!s.startsWith("/api/")}),new e.NetworkFirst({cacheName:"apis",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:16,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute((({url:e})=>{if(!(self.origin===e.origin))return!1;return!e.pathname.startsWith("/api/")}),new e.NetworkFirst({cacheName:"others",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute((({url:e})=>!(self.origin===e.origin)),new e.NetworkFirst({cacheName:"cross-origin",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:3600}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET")}));
